requirements.txt
# RAN AI Lab - Requirements
# Optimized: CPU-only torch, lightweight deps
numpy>=1.24,<2.0
pandas>=1.5,<2.1
scikit-learn>=1.2
lightgbm>=3.3
# CPU-only torch - significantly smaller than GPU wheel
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0,<2.4
onnx>=1.14
onnxruntime>=1.15
lime>=0.2
shap>=0.42
fastapi>=0.100
uvicorn[standard]>=0.23
streamlit>=1.25
psutil>=5.9
pyyaml>=6.0
pytest>=7.4
requests>=2.31
pyarrow>=12.0
